x/conflict/genesis_test.go
package conflict_test

import (
	"testing"

	keepertest "github.com/lavanet/lava/v2/testutil/keeper"
	"github.com/lavanet/lava/v2/testutil/nullify"
	"github.com/lavanet/lava/v2/x/conflict"
	"github.com/lavanet/lava/v2/x/conflict/types"
	"github.com/stretchr/testify/require"
)

func TestGenesis(t *testing.T) {
	genesisState := types.GenesisState{
		Params: types.DefaultParams(),

		ConflictVoteList: []types.ConflictVote{
			{
				Index: "0",
			},
			{
				Index: "1",
			},
		},
		// this line is used by starport scaffolding # genesis/test/state
	}

	k, ctx := keepertest.ConflictKeeper(t)
	conflict.InitGenesis(ctx, *k, genesisState)
	got := conflict.ExportGenesis(ctx, *k)
	require.NotNil(t, got)

	nullify.Fill(&genesisState)
	nullify.Fill(got)

	require.ElementsMatch(t, genesisState.ConflictVoteList, got.ConflictVoteList)
	// this line is used by starport scaffolding # genesis/test/assert
}
